+++ src/main/java/com/example/minimaltravel/dto/TransactionResponseDTO.java
@@ -5,10 +5,11 @@ import java.util.List;
 
 import com.example.minimaltravel.model.Transaction.TransactionCategory;
 
+// Envía datos de transacciones al frontend
 public class TransactionResponseDTO {
     private Long transactionId;
     private String description;
-    private Long amount;
+    private Double amount;
     private TransactionCategory category;
     private Date creationDate;
     private Long creditorUserId;
@@ -35,8 +36,8 @@ public class TransactionResponseDTO {
     public void setTransactionId(Long transactionId) { this.transactionId = transactionId; }
     public String getDescription() { return description; }
     public void setDescription(String description) { this.description = description; }
-    public Long getAmount() { return amount; }
-    public void setAmount(Long amount) { this.amount = amount; }
+    public Double getAmount() { return amount; }
+    public void setAmount(Double amount) { this.amount = amount; }
     public TransactionCategory getCategory() { return category; }
     public void setCategory(TransactionCategory category) { this.category = category; }
     public Date getCreationDate() { return creationDate; }
